chore: clean up paused statuses and move revertTrialExpiry
