+++ server/src/internal/billing/v2/setup/setupFullCustomerContext.ts
@@ -1,6 +1,15 @@
-import { EntityNotFoundError, RELEVANT_STATUSES } from "@autumn/shared";
+import {
+	CusProductStatus,
+	EntityNotFoundError,
+	RELEVANT_STATUSES,
+} from "@autumn/shared";
 import type { AutumnContext } from "@server/honoUtils/HonoEnv";
 import { CusService } from "@/internal/customers/CusService";
+
+const BILLING_CONTEXT_STATUSES = [
+	...RELEVANT_STATUSES,
+	CusProductStatus.Paused,
+];
 
 export const setupFullCustomerContext = async ({
 	ctx,
@@ -14,7 +23,7 @@ export const setupFullCustomerContext = async ({
 	const fullCustomer = await CusService.getFull({
 		ctx,
 		idOrInternalId: customerId,
-		inStatuses: RELEVANT_STATUSES,
+		inStatuses: BILLING_CONTEXT_STATUSES,
 		withEntities: true,
 		withSubs: true,
 		entityId: params.entity_id ?? undefined,
