+++ shared/utils/cusProductUtils/cusProductConstants.ts
@@ -9,7 +9,6 @@ export const RELEVANT_STATUSES = [
 	CusProductStatus.Active,
 	CusProductStatus.PastDue,
 	CusProductStatus.Scheduled,
-	CusProductStatus.Paused,
 ];
 
 export const ALL_STATUSES = [
